extends exception class to handle error code
the returned RateError has a method to access the error code and
the original response returned by fedex. this makes it easier to
map errors in a application that uses the fedex gem.

the solution is a quickfix and contains a lot of code duplication. before
merging this into the master it should be properly refactored.